+++ packages/icon/src/preload.ts
@@ -1,10 +1,7 @@
 import {fetch} from '@alwatr/fetch';
-import {contextConsumer} from '@alwatr/signal';
-
-const urlPrefixContext = contextConsumer.bind<string>('icon-url-prefix');
 
 export async function preloadIcon(name: string): Promise<string> {
-  const urlPrefix = urlPrefixContext.getValue() ?? 'https://cdn.jsdelivr.net/npm/@alwatr/icon@0/svg/';
+  const urlPrefix = 'https://cdn.jsdelivr.net/npm/@alwatr/icon@1/svg/';
   const url = urlPrefix + name + '.svg';
 
   const response = await fetch({