+++ packages/style/lib/tailwind.ts
@@ -13,7 +13,7 @@ export const tailwindConfig: Config = {
   content: [
     './res/*.html',
     './src/**/*.ts',
-    join(dirname(require.resolve('@alwatr/ui-kit2')), '**/*.ts'),
+    join(dirname(require.resolve('@alwatr/ui-kit')), '**/*.ts'),
   ],
   darkMode: 'media',
   theme: {